# project-controller/src/main/resources/static/swagger/index.yaml
swagger: '2.0'
info:
  description: new-project管理后台API
  version: 1.0.0
  title: 新项目管理后台
basePath: /
schemes:
  - http

#认证
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "token"
    in: "header"

#定义接口数据
paths:
...
  /sys/user/list:
    get:
      tags:
        - 用户管理
      summary: 查询用户列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: userName
          description: 用户名
          in: query
          type: string
        - name: phone
          description: 用户手机号码
          in: query
          type: string
        - name: deptId
          description: 部门ID
          in: query
          type: integer
        - name: status
          description: 状态 0.禁用 1.正常
          in: query
          type: integer
      responses:
        '200':
          description: 返回用户列表
          schema:
            $ref: '#/definitions/SysUserEntities'
  /sys/user/save:
    post:
      tags:
        - 用户管理
      summary: 新增用户
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户对象
          in: body
          schema:
            $ref: '#/definitions/SysUserSaveVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /sys/user/info/{userId}:
    get:
      tags:
        - 用户管理
      summary: 获取用户信息
      produces:
        - application/json
      parameters:
        - name: userId
          description: 用户ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回用户信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              user:
                $ref: '#/definitions/SysUserInfoVo'
  /sys/user/update:
    post:
      tags:
        - 用户管理
      summary: 修改用户
      produces:
        - application/json
      parameters:
        - name: body
          description: 用户对象
          in: body
          schema:
            $ref: '#/definitions/SysUserUpdateVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
...
  /deal/user/list:
    get:
      tags:
        - 客户商品-客户模块
      summary: 分页查询状态为正常的客户列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: dealUserName
          description: 客户名称
          in: query
          type: string
        - name: phone
          description: 客户手机号码
          in: query
          type: string
        - name: type
          description: 客户类型 0.个人客户 1.企业客户
          in: query
          type: integer
        - name: status
          description: 状态 0.禁用 1正常
          in: query
          type: integer
        - name: startTime
          description: 开始日期
          in: query
          type: string
          format: 'date-time'
        - name: endTime
          description: 结束日期
          in: query
          type: string
          format: 'date-time'
      responses:
        '200':
          description: 返回客户列表
          schema:
            $ref: '#/definitions/DealUserEntities'
  /deal/user/save:
    post:
      tags:
        - 客户商品-客户模块
      summary: 新增客户
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户对象
          in: body
          schema:
            $ref: '#/definitions/DealUserSaveVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /deal/user/info/{dealUserId}:
    get:
      tags:
        - 客户商品-客户模块
      summary: 获取客户信息
      produces:
        - application/json
      parameters:
        - name: dealUserId
          description: 客户ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回客户信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              config:
                description: 返回参数信息
                $ref: '#/definitions/DealUserInfoVo'
  /deal/user/update:
    post:
      tags:
        - 客户商品-客户模块
      summary: 修改客户配置
      produces:
        - application/json
      parameters:
        - name: body
          description: 客户对象
          in: body
          schema:
            $ref: '#/definitions/DealUserUpdateVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
...
  /deal/user/store/save:
    post:
      tags:
        - 客户商品-客户模块
      summary: 申请企业认证
      produces:
        - application/json
      parameters:
        - name: body
          description: 企业认证对象
          in: body
          schema:
            $ref: '#/definitions/DealUserStoreSaveVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /deal/user/store/list:
    get:
      tags:
        - 客户商品-客户模块
      summary: 分页客户查看申请成为企业的申请记录
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: dealUserId
          description: 客户ID
          in: query
          type: integer
          required: true
      responses:
        '200':
          description: 返回客户列表
          schema:
            $ref: '#/definitions/DealUserStoreEntities'
  /deal/user/store/info/{dealStoreId}:
    get:
      tags:
        - 客户商品-客户模块
      summary: 查看客户企业验证详情
      produces:
        - application/json
      parameters:
        - name: dealStoreId
          description: 客户企业ID
          in: path
          type: integer
          required: true
      responses:
        '200':
          description: 返回企业认证信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              config:
                description: 返回参数信息
                $ref: '#/definitions/DealUserStoreInfoVo'
...
  /deal/user/store/deposit/list:
    get:
      tags:
        - 客户商品-保证金模块
      summary: 分页查询保证金单列表
      produces:
        - application/json
      parameters:
        - name: page
          description: 页码
          in: query
          type: integer
          required: true
        - name: limit
          description: 每页条数
          in: query
          type: integer
          required: true
        - name: sidx
          description: 排序字段
          in: query
          type: string
        - name: order
          description: 排序方式，如：asc、desc
          in: query
          type: string
        - name: dealPhone
          description: 所属客户手机号码
          in: query
          type: string
        - name: status
          description: 审核状态 0.放弃 1.驳回 2.财务审核中 3.经理审核中 4.通过
          in: query
          type: integer
        - name: startTime
          description: 开始日期
          in: query
          type: string
          format: 'date-time'
        - name: endTime
          description: 结束日期
          in: query
          type: string
          format: 'date-time'
      responses:
        '200':
          description: 返回保证金列表
          schema:
            $ref: '#/definitions/DealUserStoreDepositEntities'
  /deal/user/store/deposit/save:
    post:
      tags:
        - 客户商品-保证金模块
      summary: 新增保证金单
      produces:
        - application/json
      parameters:
        - name: deposit
          description: 保证金对象
          in: body
          schema:
            $ref: '#/definitions/DealUserStoreDepositSaveVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /deal/user/store/deposit/info/{depositId}:
    get:
      tags:
        - 客户商品-保证金模块
      summary: 查看客户企业保证金单详情
      produces:
        - application/json
      parameters:
        - name: depositId
          description: 保证金ID
          in: path
          type: string
          required: true
      responses:
        '200':
          description: 返回保证金单信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              config:
                description: 返回参数信息
                $ref: '#/definitions/DealUserStoreDepositInfoVo'
  /deal/user/store/deposit/update:
    post:
      tags:
        - 客户商品-保证金模块
      summary: 修改保证金单
      produces:
        - application/json
      parameters:
        - name: deposit
          description: 保证金对象
          in: body
          schema:
            $ref: '#/definitions/DealUserStoreDepositUpdateVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /deal/user/store/deposit/cancel/{depositId}:
    get:
      tags:
        - 客户商品-保证金模块
      summary: 修改单据状态为放弃
      produces:
        - application/json
      parameters:
        - name: depositId
          description: 保证金ID
          in: path
          type: string
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /deal/user/store/deposit/reject:
    post:
      tags:
        - 客户商品-保证金模块
      summary: 修改单据状态为驳回
      produces:
        - application/json
      parameters:
        - name: deposit
          description: 保证金对象
          in: body
          schema:
            $ref: '#/definitions/DealUserStoreDepositInvokingVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /deal/user/store/deposit/manager:
    post:
      tags:
        - 客户商品-保证金模块
      summary: 修改单据状态为经理审核
      produces:
        - application/json
      parameters:
        - name: deposit
          description: 保证金对象
          in: body
          schema:
            $ref: '#/definitions/DealUserStoreDepositInvokingVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /deal/user/store/deposit/success:
    post:
      tags:
        - 客户商品-保证金模块
      summary: 修改单据状态为已通过
      produces:
        - application/json
      parameters:
        - name: deposit
          description: 保证金对象
          in: body
          schema:
            $ref: '#/definitions/DealUserStoreDepositInvokingVo'
          required: true
      responses:
        '200':
          description: 返回信息
          schema:
            $ref: '#/definitions/R'
  /deal/user/store/deposit/getExamineList/{depositId}:
    get:
      tags:
        - 客户商品-保证金模块
      summary: 获取保证金单据的审核记录
      produces:
        - application/json
      parameters:
        - name: depositId
          description: 保证金ID
          in: path
          type: string
          required: true
      responses:
        '200':
          description: 返回审核信息
          schema:
            type: object
            properties:
              code:
                description: 状态码 0：成功  非0：失败
                type: integer
                format: int32
              config:
                description: 返回参数信息
                $ref: '#/definitions/DealBillExamineInvokingVo'
...
definitions:
  R:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
...
  SysUserListVo:
    type: object
    properties:
      userId:
        description: 用户ID
        type: integer
        format: int64
      userName:
        description: 用户名
        type: string
      phone:
        description: 手机号
        type: string
      status:
        description: 状态  0：禁用   1：正常
        type: integer
        format: int32
      deptId:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
        format: int64
      roleList:
        description: 角色列表
        type: object
        properties:
          roleId:
            description: 角色ID
            type: integer
          roleName:
            description: 角色名称
            type: string
  SysUserEntities:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/SysUserListVo'
  SysUserInfoVo:
    type: object
    properties:
      userId:
        description: 用户ID
        type: integer
        format: int64
      userName:
        description: 用户名
        type: string
      phone:
        description: 手机号
        type: string
      passworde:
        description: 密码
        type: string
      confirm:
        description: 确认密码
        type: string
      status:
        description: 状态  0：禁用   1：正常
        type: integer
        format: int32
      deptId:
        description: 部门ID
        type: integer
        format: int64
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int32
  SysUserSaveVo:
    type: object
    properties:
      userName:
        description: 用户名
        type: string
      phone:
        description: 手机号
        type: string
      password:
        description: 密码
        type: string
      confirm:
        description: 确认密码
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int32
  SysUserUpdateVo:
    type: object
    properties:
      userId:
        description: 用户ID
        type: integer
        format: int64
      userName:
        description: 用户名
        type: string
      phone:
        description: 手机号
        type: string
      password:
        description: 密码
        type: string
      confirm:
        description: 确认密码
        type: string
      deptId:
        description: 部门ID
        type: integer
        format: int64
      roleIdList:
        description: 角色ID列表
        type: array
        items:
          type: integer
          format: int32
...
  DealUserListVo:
    type: object
    properties:
      dealUserId:
        description: 客户ID
        type: integer
        format: int64
      dealUserName:
        description: 客户名称
        type: string
      phone:
        description: 客户手机号码
        type: string
      status:
        description: 状态 0.禁用 1.正常
        type: integer
        format: int64
      type:
        description: 客户类型 0.个人客户 1.企业客户
        type: integer
        format: int64
      depositPrice:
        description: 保证金总金额
        type: string
      creditGrade:
        description: 信用等级
        type: integer
      integral:
        description: 积分
        type: integer
        format: int64
      dealStoreName:
        description: 企业名称
        type: string
      sysUserId:
        description: 所属用户ID
        type: integer
        format: int64
      sysUserName:
        description: 所属用户名称
        type: string
  DealUserEntities:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/DealUserListVo'
  DealUserInfoVo:
    type: object
    properties:
      dealUserId:
        description: 客户ID
        type: integer
        format: int64
      dealUserName:
        description: 客户名称
        type: string
      phone:
        description: 客户手机号码
        type: string
      type:
        description: 客户类型 0.个人客户 1.企业客户
        type: integer
        format: int64
      depositPrice:
        description: 保证金总金额
        type: string
      creditGrade:
        description: 信用等级
        type: integer
      integral:
        description: 积分
        type: integer
        format: int64
      dealStoreName:
        description: 所属公司
        type: string
      sysUserId:
        description: 所属用户ID
        type: integer
        format: int64
      sysUserName:
        description: 所属用户名称
        type: string
  DealUserSaveVo:
    type: object
    properties:
      dealUserName:
        description: 客户名称
        type: string
      phone:
        description: 客户手机号码
        type: string
  DealUserUpdateVo:
    type: object
    properties:
      dealUserId:
        description: 客户ID
        type: integer
        format: int64
      dealUserName:
        description: 客户名称
        type: string
...
  DealUserStoreListVo:
    type: object
    properties:
      dealStoreId:
        description: 客户企业ID
        type: integer
      dealStoreName:
        description: 客户企业名称
        type: string
      dealUserJob:
        description: 客户职位
        type: string
      sysUserId:
        description: 所属用户ID
        type: integer
        format: int64
      sysUserName:
        description: 所属用户名称
        type: string
      examineUserId:
        description: 审核人ID
        type: integer
        format: int64
      examineTime:
        description: 审核时间
        type: string
        format: date-time
  DealUserStoreEntities:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/DealUserStoreListVo'
  DealUserStoreSaveVo:
    type: object
    properties:
      dealUserId:
        description: 客户ID
        type: integer
        format: int64
      dealStoreName:
        description: 客户所属公司名称
        type: string
      image:
        description: 企业门面图路径
        type: string
      dealUserJob:
        description: 客户职位
        type: string
      sysUserId:
        description: 所属用户ID
        type: integer
        format: int64
  DealUserStoreInfoVo:
    type: object
    properties:
      dealUserId:
        description: 客户ID
        type: integer
        format: int64
      dealUserName:
        description: 客户名称
        type: string
      phone:
        description: 客户手机号码
        type: string
      type:
        description: 客户类型 0.个人客户 1.企业客户
        type: integer
        format: int64
      dealStoreName:
        description: 所属公司
        type: string
      sysUserId:
        description: 所属用户ID
        type: integer
        format: int64
      sysUserName:
        description: 所属用户名称
        type: string
      examineUserId:
        description: 审核人ID
        type: integer
        format: int64
      examineTime:
        description: 审核时间
        type: string
        format: date-time
  DealUserStoreDepositListVo:
    type: object
    properties:
      depositId:
        description: 保证金单ID
        type: string
      depositNo:
        description: 保证金单编号
        type: string
      dealStoreId:
        description: 客户企业表ID
        type: integer
        format: int64
      dealUserId:
        description: 客户ID
        type: integer
        format: int64
      dealUserName:
        description: 客户名称
        type: string
      depositPrice:
        description: 保证金金额
        type: string
      remark:
        description: 备注
        type: string
      status:
        description: 状态 0.放弃 1.驳回 2.财务审核中 3.经理审核中 4.通过
        type: integer
        format: int64
      submitTime:
        description: 提交时间
        type: string
        format: date-time
      examineId:
        description: 审核表ID
        type: integer
        format: int64
      examineUserId:
        description: 审核人ID
        type: integer
        format: int64
      examineUserName:
        description: 审核人名称
        type: string
      examineTime:
        description: 审核时间
        type: string
        format: date-time
  DealUserStoreDepositEntities:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      page:
        type: object
        properties:
          totalCount:
            description: 总记录数
            type: integer
            format: int32
          pageSize:
            description: 每页记录数
            type: integer
            format: int32
          totalPage:
            description: 总页数
            type: integer
            format: int32
          currPage:
            description: 当前页数
            type: integer
            format: int32
          list:
            type: array
            items:
              $ref: '#/definitions/DealUserStoreDepositListVo'
  DealUserStoreDepositSaveVo:
    type: object
    properties:
      dealStoreId:
        description: 客户企业表ID
        type: integer
        format: int64
      depositPrice:
        description: 保证金金额
        type: string
      remark:
        description: 备注
        type: string
  DealUserStoreDepositInfoVo:
    type: object
    properties:
      depositId:
        description: 保证金单ID
        type: string
      depositNo:
        description: 保证金单编号
        type: string
      dealStoreId:
        description: 客户企业表ID
        type: integer
        format: int64g
      dealUserId:
        description: 客户ID
        type: integer
        format: int64
      dealUserName:
        description: 客户名称
        type: string
      depositPrice:
        description: 保证金金额
        type: string
      remark:
        description: 备注
        type: string
      status:
        description: 状态 0.放弃 1.驳回 2.财务审核中 3.经理审核中 4.通过
        type: integer
        format: int64
      submitTime:
        description: 提交时间
        type: string
        format: date-time
      examine:
        type: object
        properties:
          examineId:
            description: 审核表ID
            type: integer
            format: int64
          examineUserId:
            description: 审核人ID
            type: integer
            format: int64
          examineUserName:
            description: 审核人名称
            type: string
          examineRemark:
            description: 审核理由
            type: string
          examineTime:
            description: 审核时间
            type: string
            format: date-time
  DealUserStoreDepositUpdateVo:
    type: object
    properties:
      depositId:
        description: 保证金单ID
        type: string
      depositPrice:
        description: 保证金金额
        type: string
      remark:
        description: 备注
        type: string
  DealUserStoreDepositInvokingVo:
    type: object
    properties:
      depositId:
        description: 保证金单ID
        type: string
      remark:
        description: 审核理由
        type: string
...
  DealBillExamineInvokingVo:
    type: object
    properties:
      examineId:
        description: 审核表ID
        type: string
      examineUserId:
        description: 审核人ID
        type: string
      examineUserName:
        description: 审核人名称
        type: string
      examineRemark:
        description: 审核理由
        type: string
      examineTime:
        description: 审核时间
        type: string
        format: date-time
...